@@ -9,21 +9,24 @@
         { "name": "Jonathan H. Wage", "email": "jonwage@gmail.com" }
     ],
     "require": {
-        "php": "^7.2",
-        "doctrine/collections": "~1.5",
-        "doctrine/common": "~2.9",
-        "doctrine/instantiator": "^1.1"
+        "php": "^8.0",
+        "doctrine/collections": "^1.8",
+        "doctrine/instantiator": "^1.5",
+        "doctrine/persistence": "^3.1"
     },
     "require-dev": {
-        "doctrine/coding-standard": "^8.2",
-        "phpstan/phpstan": "^0.12",
-        "phpstan/phpstan-deprecation-rules": "^0.12",
-        "phpstan/phpstan-phpunit": "^0.12",
-        "phpstan/phpstan-strict-rules": "^0.12",
-        "phpunit/phpunit": "~8.0 || ~9.0"
+        "doctrine/coding-standard": "^11.0",
+        "phpstan/phpstan": "^1.9",
+        "phpstan/phpstan-deprecation-rules": "^1.1",
+        "phpstan/phpstan-phpunit": "^1.3",
+        "phpstan/phpstan-strict-rules": "^1.4",
+        "phpunit/phpunit": "~9.5"
     },
     "config": {
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "dealerdirect/phpcodesniffer-composer-installer": true
+        }
     },
     "autoload": {
         "psr-0": {
