@@ -13,24 +13,18 @@ use function usort;
 
 class DataSourceObjectDataRepository extends BasicObjectDataRepository
 {
-    /** @var DataSource */
-    private $dataSource;
-
     /** @var mixed[][]|null */
-    private $sourceRows;
+    private array|null $sourceRows = null;
 
     public function __construct(
         ObjectManagerInterface $objectManager,
-        DataSource $dataSource,
-        string $className
+        private DataSource $dataSource,
+        string $className,
     ) {
         parent::__construct($objectManager, $className);
-        $this->dataSource = $dataSource;
     }
 
-    /**
-     * @return mixed[][]
-     */
+    /** @return mixed[][] */
     public function findAll(): array
     {
         return $this->getSourceRows();
@@ -44,9 +38,9 @@ class DataSourceObjectDataRepository extends BasicObjectDataRepository
      */
     public function findBy(
         array $criteria,
-        ?array $orderBy = null,
-        ?int $limit = null,
-        ?int $offset = null
+        array|null $orderBy = null,
+        int|null $limit = null,
+        int|null $offset = null,
     ): array {
         $rows = [];
 
@@ -74,7 +68,7 @@ class DataSourceObjectDataRepository extends BasicObjectDataRepository
      *
      * @return mixed[]|null
      */
-    public function findOneBy(array $criteria): ?array
+    public function findOneBy(array $criteria): array|null
     {
         foreach ($this->getSourceRows() as $row) {
             if ($this->matches($criteria, $row)) {
@@ -112,7 +106,7 @@ class DataSourceObjectDataRepository extends BasicObjectDataRepository
      *
      * @return mixed[][] $rows
      */
-    private function slice(array $rows, ?int $limit, ?int $offset): array
+    private function slice(array $rows, int|null $limit, int|null $offset): array
     {
         if ($limit === null) {
             $limit = count($rows);
@@ -125,9 +119,7 @@ class DataSourceObjectDataRepository extends BasicObjectDataRepository
         return array_slice($rows, $offset, $limit);
     }
 
-    /**
-     * @return mixed[][]
-     */
+    /** @return mixed[][] */
     private function getSourceRows(): array
     {
         if ($this->sourceRows === null) {
