@@ -13,22 +13,12 @@ use Doctrine\SkeletonMapper\Event\PreUpdateEventArgs;
 use Doctrine\SkeletonMapper\Events;
 use Doctrine\SkeletonMapper\ObjectManagerInterface;
 
-use function get_class;
-
 class EventDispatcher
 {
-    /** @var ObjectManagerInterface */
-    private $objectManager;
-
-    /** @var EventManager */
-    private $eventManager;
-
     public function __construct(
-        ObjectManagerInterface $objectManager,
-        EventManager $eventManager
+        private ObjectManagerInterface $objectManager,
+        private EventManager $eventManager,
     ) {
-        $this->objectManager = $objectManager;
-        $this->eventManager  = $eventManager;
     }
 
     public function dispatchEvent(string $eventName, EventArgs $event): void
@@ -40,13 +30,10 @@ class EventDispatcher
         $this->eventManager->dispatchEvent($eventName, $event);
     }
 
-    /**
-     * @param object  $object
-     * @param mixed[] $args
-     */
-    public function dispatchObjectLifecycleCallback(string $eventName, $object, array &$args = []): void
+    /** @param mixed[] $args */
+    public function dispatchObjectLifecycleCallback(string $eventName, object $object, array &$args = []): void
     {
-        $className = get_class($object);
+        $className = $object::class;
 
         $class = $this->objectManager->getClassMetadata($className);
 
@@ -57,9 +44,7 @@ class EventDispatcher
         $class->invokeLifecycleCallbacks($eventName, $object, $args);
     }
 
-    /**
-     * @param object[] $objects
-     */
+    /** @param object[] $objects */
     public function dispatchObjectsLifecycleCallbacks(string $eventName, array $objects): void
     {
         foreach ($objects as $object) {
@@ -71,13 +56,11 @@ class EventDispatcher
     {
         $this->dispatchEvent(
             Events::preFlush,
-            new Event\PreFlushEventArgs($this->objectManager)
+            new Event\PreFlushEventArgs($this->objectManager),
         );
     }
 
-    /**
-     * @param object[] $objects
-     */
+    /** @param object[] $objects */
     public function dispatchPreFlushLifecycleCallbacks(array $objects): void
     {
         $this->dispatchObjectsLifecycleCallbacks(Events::preFlush, $objects);
@@ -87,7 +70,7 @@ class EventDispatcher
     {
         $this->dispatchEvent(
             Events::onFlush,
-            new Event\OnFlushEventArgs($this->objectManager)
+            new Event\OnFlushEventArgs($this->objectManager),
         );
     }
 
@@ -95,35 +78,29 @@ class EventDispatcher
     {
         $this->dispatchEvent(
             Events::postFlush,
-            new Event\PostFlushEventArgs($this->objectManager)
+            new Event\PostFlushEventArgs($this->objectManager),
         );
     }
 
-    public function dispatchOnClearEvent(?string $className): void
+    public function dispatchOnClearEvent(string|null $className): void
     {
         $this->dispatchEvent(
             Events::onClear,
-            new Event\OnClearEventArgs($this->objectManager, $className)
+            new Event\OnClearEventArgs($this->objectManager),
         );
     }
 
-    /**
-     * @param object $object
-     */
-    public function dispatchPreRemove($object): void
+    public function dispatchPreRemove(object $object): void
     {
         $this->dispatchObjectLifecycleCallback(Events::preRemove, $object);
 
         $this->dispatchEvent(
             Events::preRemove,
-            new LifecycleEventArgs($object, $this->objectManager)
+            new LifecycleEventArgs($object, $this->objectManager),
         );
     }
 
-    /**
-     * @param object $object
-     */
-    public function dispatchPreUpdate($object, ChangeSet $changeSet): void
+    public function dispatchPreUpdate(object $object, ChangeSet $changeSet): void
     {
         $args = [$changeSet];
         $this->dispatchObjectLifecycleCallback(Events::preUpdate, $object, $args);
@@ -133,57 +110,45 @@ class EventDispatcher
             new PreUpdateEventArgs(
                 $object,
                 $this->objectManager,
-                $changeSet
-            )
+                $changeSet,
+            ),
         );
     }
 
-    /**
-     * @param object $object
-     */
-    public function dispatchPrePersist($object): void
+    public function dispatchPrePersist(object $object): void
     {
         $this->dispatchObjectLifecycleCallback(Events::prePersist, $object);
 
         $this->dispatchEvent(
             Events::prePersist,
-            new LifecycleEventArgs($object, $this->objectManager)
+            new LifecycleEventArgs($object, $this->objectManager),
         );
     }
 
-    /**
-     * @param object  $object
-     * @param mixed[] $data
-     */
-    public function dispatchPreLoad($object, array &$data): void
+    /** @param mixed[] $data */
+    public function dispatchPreLoad(object $object, array &$data): void
     {
         $args = [&$data];
         $this->dispatchObjectLifecycleCallback(Events::preLoad, $object, $args);
 
         $this->dispatchEvent(
             Events::preLoad,
-            new PreLoadEventArgs($object, $this->objectManager, $data)
+            new PreLoadEventArgs($object, $this->objectManager, $data),
         );
     }
 
-    /**
-     * @param object $object
-     */
-    public function dispatchPostLoad($object): void
+    public function dispatchPostLoad(object $object): void
     {
         $this->dispatchLifecycleEvent(Events::postLoad, $object);
     }
 
-    /**
-     * @param object $object
-     */
-    public function dispatchLifecycleEvent(string $eventName, $object): void
+    public function dispatchLifecycleEvent(string $eventName, object $object): void
     {
         $this->dispatchObjectLifecycleCallback($eventName, $object);
 
         $this->dispatchEvent(
             $eventName,
-            new LifecycleEventArgs($object, $this->objectManager)
+            new LifecycleEventArgs($object, $this->objectManager),
         );
     }
 }
