@@ -9,16 +9,12 @@ use Doctrine\SkeletonMapper\Hydrator\HydratableInterface;
 use Doctrine\SkeletonMapper\ObjectManagerInterface;
 use PHPUnit\Framework\TestCase;
 
-/**
- * @group unit
- */
+/** @group unit */
 class BasicObjectHydratorTest extends TestCase
 {
-    /** @var ObjectManagerInterface */
-    private $objectManager;
+    private ObjectManagerInterface $objectManager;
 
-    /** @var BasicObjectHydrator */
-    private $hydrator;
+    private BasicObjectHydrator $hydrator;
 
     public function testHydrate(): void
     {
@@ -41,11 +37,9 @@ class BasicObjectHydratorTest extends TestCase
 class HydratableObject implements HydratableInterface
 {
     /** @var mixed[] */
-    public $data;
+    public array $data;
 
-    /**
-     * @param mixed[] $data
-     */
+    /** @param mixed[] $data */
     public function hydrate(array $data, ObjectManagerInterface $objectManager): void
     {
         $this->data = $data;
