@@ -5,19 +5,16 @@ declare(strict_types=1);
 namespace Doctrine\SkeletonMapper\Tests\Mapping;
 
 use BadMethodCallException;
-use Doctrine\Common\Persistence\Event\LifecycleEventArgs;
 use Doctrine\SkeletonMapper\Mapping\ClassMetadata;
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
 use stdClass;
 
-/**
- * @group unit
- */
+/** @group unit */
 class ClassMetadataTest extends TestCase
 {
-    /** @var ClassMetadata */
-    private $class;
+    /** @var ClassMetadata<object> */
+    private ClassMetadata $class;
 
     public function testMapField(): void
     {
@@ -133,7 +130,7 @@ class ClassMetadataTest extends TestCase
                 'fieldName' => 'groups',
                 'targetObject' => 'Test',
                 'type' => 'many',
-            ]
+            ],
         );
 
         self::assertFalse($this->class->hasField('groups'));
@@ -204,7 +201,6 @@ class ClassMetadataTest extends TestCase
     public function testInvokeLifecycleCallbacksWithoutArguments(): void
     {
         $object = new ClassMetadataTestModel();
-        $data   = ['test'];
 
         $this->class->lifecycleCallbacks['test'] = ['testEvent'];
 
@@ -281,19 +277,15 @@ class ClassMetadataTest extends TestCase
 
 class ClassMetadataTestModel
 {
-    /** @var int */
-    public $id;
+    public int $id;
 
-    /** @var string */
-    public $name;
+    public string $name;
 
-    /** @var LifecycleEventArgs|true */
-    public $testEventCalled;
+    /** @var array<string>|true */
+    public array|bool $testEventCalled;
 
-    /**
-     * @param LifecycleEventArgs|true $args
-     */
-    public function testEvent($args = null): void
+    /** @param array<string>|true $args */
+    public function testEvent(array|bool|null $args = null): void
     {
         if ($args !== null) {
             $this->testEventCalled = $args;
