@@ -9,13 +9,10 @@ use Doctrine\SkeletonMapper\ObjectRepository\ObjectRepositoryInterface;
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
 
-/**
- * @group unit
- */
+/** @group unit */
 class ObjectRepositoryFactoryTest extends TestCase
 {
-    /** @var ObjectRepositoryFactory */
-    private $factory;
+    private ObjectRepositoryFactory $factory;
 
     public function testAddObjectRepository(): void
     {
