@@ -17,31 +17,24 @@ use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use stdClass;
 
-/**
- * @group unit
- */
+/** @group unit */
 class ObjectRepositoryTest extends TestCase
 {
-    /** @var ObjectManagerInterface|MockObject */
-    private $objectManager;
+    private ObjectManagerInterface|MockObject $objectManager;
 
-    /** @var ObjectDataRepositoryInterface|MockObject */
-    private $objectDataRepository;
+    private ObjectDataRepositoryInterface|MockObject $objectDataRepository;
 
-    /** @var ObjectFactory|MockObject */
-    private $objectFactory;
+    private ObjectFactory|MockObject $objectFactory;
 
-    /** @var ObjectHydratorInterface|MockObject */
-    private $hydrator;
+    private ObjectHydratorInterface|MockObject $hydrator;
 
-    /** @var EventManager|MockObject */
-    private $eventManager;
+    private EventManager|MockObject $eventManager;
 
-    /** @var ClassMetadataInterface|MockObject */
+    /** @phpstan-var ClassMetadataInterface<object>|MockObject */
     private $classMetadata;
 
-    /** @var TestObjectRepository */
-    private $repository;
+    /** @var TestObjectRepository<object> */
+    private TestObjectRepository $repository;
 
     public function testGetClassName(): void
     {
@@ -191,24 +184,25 @@ class ObjectRepositoryTest extends TestCase
             $this->objectFactory,
             $this->hydrator,
             $this->eventManager,
-            ArrayObject::class
+            ArrayObject::class,
         );
     }
 }
 
+/**
+ * @template T of object
+ * @template-implements ObjectRepository<T>
+ */
 class TestObjectRepository extends ObjectRepository
 {
+    /** @return ClassMetadataInterface<object> */
     public function getClassMetadata(): ClassMetadataInterface
     {
         return $this->class;
     }
 
-    /**
-     * @param object $object
-     *
-     * @return int[]
-     */
-    public function getObjectIdentifier($object): array
+    /** @return int[] */
+    public function getObjectIdentifier(object $object): array
     {
         return ['id' => 1];
     }
@@ -223,10 +217,7 @@ class TestObjectRepository extends ObjectRepository
         return ['id' => 1];
     }
 
-    /**
-     * @param object $object
-     */
-    public function merge($object): void
+    public function merge(object $object): void
     {
     }
 }
