@@ -14,19 +14,15 @@ use Doctrine\SkeletonMapper\UnitOfWork\ChangeSet;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 
-/**
- * @group unit
- */
+/** @group unit */
 class ArrayObjectPersisterTest extends TestCase
 {
-    /** @var ObjectManagerInterface|MockObject */
-    private $objectManager;
+    private ObjectManagerInterface|MockObject $objectManager;
 
     /** @var ArrayCollection<int, array<string, mixed>> */
-    private $objects;
+    private ArrayCollection $objects;
 
-    /** @var ArrayObjectPersister */
-    private $persister;
+    private ArrayObjectPersister $persister;
 
     /** @phpstan-var class-string */
     private $testClassName = ArrayObjectPersisterTestModel::class;
@@ -112,24 +108,20 @@ class ArrayObjectPersisterTest extends TestCase
         $this->persister = new ArrayObjectPersister(
             $this->objectManager,
             $this->objects,
-            $this->testClassName
+            $this->testClassName,
         );
     }
 }
 
 class ArrayObjectPersisterTestModel implements PersistableInterface
 {
-    /**
-     * @return string[]
-     */
+    /** @return string[] */
     public function preparePersistChangeSet(): array
     {
         return ['username' => 'jwage'];
     }
 
-    /**
-     * @return string[]
-     */
+    /** @return string[] */
     public function prepareUpdateChangeSet(ChangeSet $changeSet): array
     {
         $changes = [];
