@@ -10,7 +10,9 @@ use Doctrine\Common\Collections\Collection;
 use function call_user_func;
 
 /**
- * @template-extends AbstractLazyCollection<mixed, mixed>
+ * @psalm-template TKey of array-key
+ * @psalm-template T
+ * @template-extends AbstractLazyCollection<TKey,T>
  */
 class LazyCollection extends AbstractLazyCollection
 {
@@ -22,9 +24,7 @@ class LazyCollection extends AbstractLazyCollection
         $this->callback = $callback;
     }
 
-    /**
-     * @return Collection<int, object>
-     */
+    /** @return Collection<TKey,T> */
     public function getCollection(): Collection
     {
         $this->initialize();
