@@ -4,11 +4,9 @@ declare(strict_types=1);
 
 namespace Doctrine\SkeletonMapper\Mapping;
 
+/** @template-implements ClassMetadataInstantiatorInterface<object> */
 class ClassMetadataInstantiator implements ClassMetadataInstantiatorInterface
 {
-    /**
-     * @param class-string $className
-     */
     public function instantiate(string $className): ClassMetadata
     {
         return new ClassMetadata($className);
