@@ -4,21 +4,22 @@ declare(strict_types=1);
 
 namespace Doctrine\SkeletonMapper\ObjectRepository;
 
-use Doctrine\Common\Persistence\ObjectRepository as BaseObjectRepositoryInterface;
+use Doctrine\Persistence\ObjectRepository as BaseObjectRepositoryInterface;
 
 /**
  * Interface that object repositories must implement.
+ *
+ * @template T of object
+ * @template-extends BaseObjectRepositoryInterface<T>
  */
 interface ObjectRepositoryInterface extends BaseObjectRepositoryInterface
 {
     /**
      * Returns the objects identifier.
      *
-     * @param object $object
-     *
      * @return mixed[]
      */
-    public function getObjectIdentifier($object): array;
+    public function getObjectIdentifier(object $object): array;
 
     /**
      * Returns the identifier.
@@ -29,26 +30,13 @@ interface ObjectRepositoryInterface extends BaseObjectRepositoryInterface
      */
     public function getObjectIdentifierFromData(array $data): array;
 
-    /**
-     * @param object $object
-     */
-    public function merge($object): void;
+    public function merge(object $object): void;
 
-    /**
-     * @param object  $object
-     * @param mixed[] $data
-     */
-    public function hydrate($object, array $data): void;
+    /** @param mixed[] $data */
+    public function hydrate(object $object, array $data): void;
 
-    /**
-     * @return object
-     *
-     * @phpstan-param class-string $className
-     */
-    public function create(string $className);
+    /** @phpstan-param class-string $className */
+    public function create(string $className): object;
 
-    /**
-     * @param object $object
-     */
-    public function refresh($object): void;
+    public function refresh(object $object): void;
 }
