@@ -11,31 +11,26 @@ use Doctrine\SkeletonMapper\UnitOfWork\ChangeSet;
 
 use function max;
 
+/**
+ * @template T of object
+ * @template-extends BasicObjectPersister<T>
+ */
 class ArrayObjectPersister extends BasicObjectPersister
 {
-    /** @var ArrayCollection<mixed, mixed> */
-    protected $objects;
-
     /**
-     * @param ArrayCollection<mixed, mixed> $objects
-     * @param class-string                  $className
+     * @param ArrayCollection<int|string, mixed> $objects
+     * @param class-string<T>                    $className
      */
     public function __construct(
         ObjectManagerInterface $objectManager,
-        ArrayCollection $objects,
-        string $className
+        protected ArrayCollection $objects,
+        string $className,
     ) {
         parent::__construct($objectManager, $className);
-
-        $this->objects = $objects;
     }
 
-    /**
-     * @param object $object
-     *
-     * @return mixed[]
-     */
-    public function persistObject($object): array
+    /** @return mixed[] */
+    public function persistObject(object $object): array
     {
         $data = $this->preparePersistChangeSet($object);
 
@@ -50,12 +45,8 @@ class ArrayObjectPersister extends BasicObjectPersister
         return $data;
     }
 
-    /**
-     * @param object $object
-     *
-     * @return mixed[]
-     */
-    public function updateObject($object, ChangeSet $changeSet): array
+    /** @return mixed[] */
+    public function updateObject(object $object, ChangeSet $changeSet): array
     {
         $changeSet = $this->prepareUpdateChangeSet($object, $changeSet);
 
@@ -73,10 +64,7 @@ class ArrayObjectPersister extends BasicObjectPersister
         return $objectData;
     }
 
-    /**
-     * @param object $object
-     */
-    public function removeObject($object): void
+    public function removeObject(object $object): void
     {
         $class      = $this->getClassMetadata();
         $identifier = $this->getObjectIdentifier($object);
@@ -84,6 +72,7 @@ class ArrayObjectPersister extends BasicObjectPersister
         unset($this->objects[$identifier[$class->getIdentifier()[0]]]);
     }
 
+    /** @phpstan-param ClassMetadataInterface<T> $class */
     private function generateNextId(ClassMetadataInterface $class): int
     {
         $ids = [];
